fix: Move studio dashboard from admin API to public API
The studio/broadcaster dashboard must be accessible on the public API
(port 9000) rather than the internal admin API (port 9001). This fixes
a critical architectural mistake that made the studio inaccessible to
broadcasters.

- Removed studio static files, routes, and router includes from admin.py
- Added studio static files and /studio route to public.py
- Added studio and studio_sse router includes to public.py
